+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - macOS and iOS building, testing and publication tasks to CI configuration
 - Linux ARM64 target support
 - Support for SHA-1 digest algorithm
+- API for Key Derivation Functions (KDFs) with support for PBKDF2, HKDF, the Argon2 family
 
 ### Changed
 - Complete restructuring of the API. Removed internal APIs in favor of modular algorithm interface.